Move exceptions to package
