+++ src/main/java/momo/Momo.java
@@ -1,6 +1,7 @@
 package momo;
 
 import momo.command.*;
+import momo.exception.MomoException;
 import momo.task.TaskList;
 
 /**
