+++ src/main/java/momo/exception/EmptyCommandException.java
@@ -1,4 +1,4 @@
-package momo;
+package momo.exception;
 
 /**
  * Thrown when user gives empty command
