+++ src/main/java/momo/exception/InvalidCommandException.java
@@ -1,4 +1,4 @@
-package momo;
+package momo.exception;
 
 /**
  * Thrown when user gives wrongly formatted command
