+++ src/test/java/momo/command/DeleteCommandTest.java
@@ -1,16 +1,12 @@
 package momo.command;
 
-import momo.InvalidCommandException;
-import momo.MomoException;
+import momo.exception.InvalidCommandException;
 import momo.StorageStub;
 import momo.task.TasklistStub;
 import momo.task.Todo;
 import momo.task.Task;
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-
-import java.time.LocalDate;
 
 import static org.junit.jupiter.api.Assertions.*;
